add round completion check to update match API, protect MatchCard against undefined player names, move tournament logic from frontend to backend using the new APIs, clean up unused tournament data  structures from frontend, update Match type to include tournamentId
